refactor(32): change imports from types folder to entities
